fix: export class in a separate file
