+++ ng2-slimscroll.metadata.json
@@ -1,1 +1,1 @@
-[{"__symbolic":"module","version":2,"metadata":{"SlimScrollModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}],"exports":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}]}]}]}},"exports":[{"from":"./src/directives/slimscroll.directive","export":["ISlimScrollOptions","SlimScrollOptions"]}]},{"__symbolic":"module","exports":[{"from":"./src/directives/slimscroll.directive","export":["ISlimScrollOptions","SlimScrollOptions"]}],"metadata":{"SlimScrollModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}],"exports":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}]}]}]}},"version":1}]
+[{"__symbolic":"module","version":2,"metadata":{"SlimScrollModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}],"exports":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}]}]}]}},"exports":[{"from":"./src/classes/slimscroll-options.class"}]},{"__symbolic":"module","exports":[{"from":"./src/classes/slimscroll-options.class"}],"metadata":{"SlimScrollModule":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"NgModule"},"arguments":[{"declarations":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}],"exports":[{"__symbolic":"reference","module":"./src/directives/slimscroll.directive","name":"SlimScrollDirective"}]}]}]}},"version":1}]
