+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng2-slimscroll",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Angular2 Customizable Scrollbar Implementation",
   "main": "ng2-datepicker.js",
   "repository": {
