+++ src/classes/slimscroll-options.class.d.ts
@@ -1,0 +1,27 @@
+export interface ISlimScrollOptions {
+    position?: string;
+    barBackground?: string;
+    barOpacity?: string;
+    barWidth?: string;
+    barBorderRadius?: string;
+    barMargin?: string;
+    gridBackground?: string;
+    gridOpacity?: string;
+    gridWidth?: string;
+    gridBorderRadius?: string;
+    gridMargin?: string;
+}
+export declare class SlimScrollOptions {
+    position?: string;
+    barBackground?: string;
+    barOpacity?: string;
+    barWidth?: string;
+    barBorderRadius?: string;
+    barMargin?: string;
+    gridBackground?: string;
+    gridOpacity?: string;
+    gridWidth?: string;
+    gridBorderRadius?: string;
+    gridMargin?: string;
+    constructor(obj?: ISlimScrollOptions);
+}
