+++ src/classes/slimscroll-options.class.js
@@ -1,0 +1,16 @@
+export var SlimScrollOptions = (function () {
+    function SlimScrollOptions(obj) {
+        this.position = obj && obj.position ? obj.position : 'right';
+        this.barBackground = obj && obj.barBackground ? obj.barBackground : '#343a40';
+        this.barOpacity = obj && obj.barOpacity ? obj.barOpacity : '1';
+        this.barWidth = obj && obj.barWidth ? obj.barWidth : '12';
+        this.barBorderRadius = obj && obj.barBorderRadius ? obj.barBorderRadius : '5';
+        this.barMargin = obj && obj.barMargin ? obj.barMargin : '1px 0';
+        this.gridBackground = obj && obj.gridBackground ? obj.gridBackground : '#adb5bd';
+        this.gridOpacity = obj && obj.gridOpacity ? obj.gridOpacity : '1';
+        this.gridWidth = obj && obj.gridWidth ? obj.gridWidth : '8';
+        this.gridBorderRadius = obj && obj.gridBorderRadius ? obj.gridBorderRadius : '10';
+        this.gridMargin = obj && obj.gridMargin ? obj.gridMargin : '1px 2px';
+    }
+    return SlimScrollOptions;
+}());
