+++ src/classes/slimscroll-options.class.metadata.json
@@ -1,0 +1,1 @@
+[{"__symbolic":"module","version":2,"metadata":{"SlimScrollOptions":{"__symbolic":"class","members":{"__ctor__":[{"__symbolic":"constructor","parameters":[{"__symbolic":"error","message":"Could not resolve type","line":27,"character":20,"context":{"typeName":"ISlimScrollOptions"}}]}]}}}},{"__symbolic":"module","metadata":{"SlimScrollOptions":{"__symbolic":"class","members":{"__ctor__":[{"__symbolic":"constructor","parameters":[{"__symbolic":"error","message":"Could not resolve type","line":27,"character":20,"context":{"typeName":"ISlimScrollOptions"}}]}]}}},"version":1}]
