+++ src/directives/slimscroll.directive.metadata.json
@@ -1,1 +1,1 @@
-[{"__symbolic":"module","version":2,"metadata":{"SlimScrollOptions":{"__symbolic":"class","members":{"__ctor__":[{"__symbolic":"constructor","parameters":[{"__symbolic":"error","message":"Could not resolve type","line":29,"character":20,"context":{"typeName":"ISlimScrollOptions"}}]}]}},"SlimScrollDirective":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Directive"},"arguments":[{"selector":"[slimScroll]"}]}],"members":{"options":[{"__symbolic":"property","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Input"}}]}],"__ctor__":[{"__symbolic":"constructor","parameterDecorators":[[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"}]}],[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}]],"parameters":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"},{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}],"ngOnInit":[{"__symbolic":"method"}],"setElementStyle":[{"__symbolic":"method"}],"onMutation":[{"__symbolic":"method"}],"wrapContainer":[{"__symbolic":"method"}],"initGrid":[{"__symbolic":"method"}],"initBar":[{"__symbolic":"method"}],"getBarHeight":[{"__symbolic":"method"}],"attachWheel":[{"__symbolic":"method"}],"scrollContent":[{"__symbolic":"method"}],"destroy":[{"__symbolic":"method"}],"unwrap":[{"__symbolic":"method"}]}}}},{"__symbolic":"module","metadata":{"SlimScrollOptions":{"__symbolic":"class","members":{"__ctor__":[{"__symbolic":"constructor","parameters":[{"__symbolic":"error","message":"Could not resolve type","line":29,"character":20,"context":{"typeName":"ISlimScrollOptions"}}]}]}},"SlimScrollDirective":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Directive"},"arguments":[{"selector":"[slimScroll]"}]}],"members":{"options":[{"__symbolic":"property","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Input"}}]}],"__ctor__":[{"__symbolic":"constructor","parameterDecorators":[[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"}]}],[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}]],"parameters":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"},{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}],"ngOnInit":[{"__symbolic":"method"}],"setElementStyle":[{"__symbolic":"method"}],"onMutation":[{"__symbolic":"method"}],"wrapContainer":[{"__symbolic":"method"}],"initGrid":[{"__symbolic":"method"}],"initBar":[{"__symbolic":"method"}],"getBarHeight":[{"__symbolic":"method"}],"attachWheel":[{"__symbolic":"method"}],"scrollContent":[{"__symbolic":"method"}],"destroy":[{"__symbolic":"method"}],"unwrap":[{"__symbolic":"method"}]}}},"version":1}]
+[{"__symbolic":"module","version":2,"metadata":{"SlimScrollDirective":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Directive"},"arguments":[{"selector":"[slimScroll]"}]}],"members":{"options":[{"__symbolic":"property","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Input"}}]}],"__ctor__":[{"__symbolic":"constructor","parameterDecorators":[[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"}]}],[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}]],"parameters":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"},{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}],"ngOnInit":[{"__symbolic":"method"}],"setElementStyle":[{"__symbolic":"method"}],"onMutation":[{"__symbolic":"method"}],"wrapContainer":[{"__symbolic":"method"}],"initGrid":[{"__symbolic":"method"}],"initBar":[{"__symbolic":"method"}],"getBarHeight":[{"__symbolic":"method"}],"attachWheel":[{"__symbolic":"method"}],"scrollContent":[{"__symbolic":"method"}],"destroy":[{"__symbolic":"method"}],"unwrap":[{"__symbolic":"method"}]}}}},{"__symbolic":"module","metadata":{"SlimScrollDirective":{"__symbolic":"class","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Directive"},"arguments":[{"selector":"[slimScroll]"}]}],"members":{"options":[{"__symbolic":"property","decorators":[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Input"}}]}],"__ctor__":[{"__symbolic":"constructor","parameterDecorators":[[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"}]}],[{"__symbolic":"call","expression":{"__symbolic":"reference","module":"@angular/core","name":"Inject"},"arguments":[{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}]],"parameters":[{"__symbolic":"reference","module":"@angular/core","name":"ViewContainerRef"},{"__symbolic":"reference","module":"@angular/core","name":"Renderer"}]}],"ngOnInit":[{"__symbolic":"method"}],"setElementStyle":[{"__symbolic":"method"}],"onMutation":[{"__symbolic":"method"}],"wrapContainer":[{"__symbolic":"method"}],"initGrid":[{"__symbolic":"method"}],"initBar":[{"__symbolic":"method"}],"getBarHeight":[{"__symbolic":"method"}],"attachWheel":[{"__symbolic":"method"}],"scrollContent":[{"__symbolic":"method"}],"destroy":[{"__symbolic":"method"}],"unwrap":[{"__symbolic":"method"}]}}},"version":1}]
